Update(Notifications): Send notifications to multiple users & Delete notifications & Pagination

+++ backend/controllers/notifications.js
@@ -2,30 +2,41 @@ const _ = require("lodash");
 const { Notification } = require("../models");
 const client = require("../caches/redis");
 
-module.exports.create = async (req, res) => {
+module.exports.send = async (req, res) => {
   try {
-    const base = {
+    const baseReceived = {
+      type: "received",
+      checked: false,
       fromUserId: req.user.userId,
       fromUserName: req.user.userName,
-      type: req.body.type,
-      message: req.body.message,
+      category: req.body.category,
+      title: req.body.title,
+      description: req.body.description,
     };
 
-    const notifications = req.body.toUsers.map((toUser) => {
-      return {
-        ...base,
-        toUserId: toUser.userId,
-        toUserName: toUser.userName,
-      };
+    const notifications = req.body.toUserList.map((user) => ({
+      ...baseReceived,
+      userId: user.userId,
+      userName: user.userName,
+    }));
+
+    notifications.push({
+      type: "sent",
+      userId: req.user.userId,
+      userName: req.user.userName,
+      toUserList: req.body.toUserList,
+      category: req.body.category,
+      title: req.body.title,
+      description: req.body.description,
     });
 
     const newNotifications = await Notification(req.user.academyId).insertMany(
       notifications
     );
 
-    for (let notification of newNotifications) {
+    for (let notification of _.filter(newNotifications, { type: "received" })) {
       await client.set(
-        `isReceivedNotifications/${req.user.academyId}/${notification.toUserId}`,
+        `isReceivedNotifications/${req.user.academyId}/${notification.userId}`,
         "true"
       );
     }
@@ -38,15 +49,42 @@ module.exports.create = async (req, res) => {
 
 module.exports.find = async (req, res) => {
   try {
-    const notifications = await Notification(req.user.academyId).find(
+    const requestPage = parseInt(req.query.page);
+    const requestSize = parseInt(req.query.size);
+    delete req.query.requestPage;
+    delete req.query.requestSize;
+
+    const totalDataCnt = await Notification(req.user.academyId).countDocuments(
       req.query
     );
 
-    await client.del(
-      `isReceivedNotifications/${req.user.academyId}/${req.query.toUserId}`
-    );
+    const totalPages = Math.ceil(totalDataCnt / requestSize);
+    const isLastPage = requestPage === totalPages;
+    const isFirstPage = requestPage === 1;
 
-    return res.status(200).send({ notifications });
+    const notifications = await Notification(req.user.academyId)
+      .find(req.query)
+      .sort({ createdAt: -1 })
+      .limit(requestSize)
+      .skip(requestSize * (requestPage - 1));
+
+    if (req.query.type === "received") {
+      await client.del(
+        `isReceivedNotifications/${req.user.academyId}/${req.query.userId}`
+      );
+    }
+
+    return res.status(200).send({
+      notifications,
+      page: {
+        totalDataCnt,
+        totalPages,
+        isLastPage,
+        isFirstPage,
+        requestPage,
+        requestSize,
+      },
+    });
   } catch (err) {
     return res.status(500).send({ message: err.message });
   }
@@ -78,13 +116,12 @@ module.exports.uncheck = async (req, res) => {
 
 module.exports.remove = async (req, res) => {
   try {
-    const notification = await Notification(req.user.academyId).findById(
-      req.params._id
-    );
-    if (!notification) return res.status(404).send({ message: "td not found" });
+    const ids = _.split(req.params._ids, "&");
+    const result = await Notification(req.user.academyId).deleteMany({
+      _id: { $in: ids },
+    });
 
-    await notification.remove();
-    return res.status(200).send();
+    return res.status(200).send(result);
   } catch (err) {
     return res.status(500).send({ message: err.message });
   }
